+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -216,6 +216,7 @@ export class GeckoViewStartup {
               "GeckoView:IPProtection:IPProtectionService:GetState",
               "GeckoView:IPProtection:IPPProxyManager:GetState",
               "GeckoView:IPProtection:Activate",
+              "GeckoView:IPProtection:Switch",
               "GeckoView:IPProtection:Deactivate",
               "GeckoView:IPProtection:Enroll",
               "GeckoView:IPProtection:RefreshUsage",