+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -22,6 +22,7 @@ exclude: true
   ([bug 2042799]({{bugzilla}}2042799))
 - Added [`IPProtectionController.getServerList`][153.7] and [`IPProtectionController.Country`][153.8] to retrieve the list of countries available in the proxy serverlist.
 - Added [`IPProtectionController.activate(Country)`][153.9] to activate the proxy routing through a selected country.
+- Added [`IPProtectionController.switchTo(Country)`][153.10] to switch the active proxy connection to a server in a different country, resolving to whether the connection was switched.
 
 [153.1]: {{javadoc_uri}}/GeckoSession.ScrollPositionUpdate.SourceType.html
 [153.2]: {{javadoc_uri}}/GeckoSession.ScrollPositionUpdate.html#source
@@ -32,6 +33,7 @@ exclude: true
 [153.7]: {{javadoc_uri}}/IPProtectionController.html#getServerList()
 [153.8]: {{javadoc_uri}}/IPProtectionController.Country.html
 [153.9]: {{javadoc_uri}}/IPProtectionController.html#activate(org.mozilla.geckoview.IPProtectionController.Country)
+[153.10]: {{javadoc_uri}}/IPProtectionController.html#switchTo(org.mozilla.geckoview.IPProtectionController.Country)
 
 ## v152
 - Added [`WebExtensionController.INSTALLATION_METHOD_RTAMO`][152.1], which should be used when a `WebExtension` is installed for the RTAMO feature. ([bug 2029607]({{bugzilla}}2029607))
@@ -1995,4 +1997,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: e32264a4f6fe764ff42cf9266faa52991d330717
+[api-version]: 763edcba86dd510a921c1a4778dae99abc854cd4